Index page minor changes
Added index page style sheet and corrected some mistakes in index.php, however the part where radio buttons are created using query results is commented out to be able to test the rest of the page. I will update that part when database is ready. I'll first correct the schema for that.